small refactor

+++ ff/src/fields/fft_friendly.rs
@@ -56,7 +56,7 @@ pub trait FftField: crate::Field {
 
             omega = large_subgroup_root_of_unity;
             for _ in q_adicity..small_subgroup_base_adicity {
-                omega = omega.pow([q as u64]);
+                omega = omega.pow([q]);
             }
 
             for _ in two_adicity..Self::TWO_ADICITY {
@@ -64,7 +64,7 @@ pub trait FftField: crate::Field {
             }
         } else {
             // Compute the next power of 2.
-            let size = n.next_power_of_two() as u64;
+            let size = n.next_power_of_two();
             let log_size_of_group = ark_std::log2(usize::try_from(size).expect("too large"));
 
             if n != size || log_size_of_group > Self::TWO_ADICITY {
